Fix merge consflict.
