+++ books.json
@@ -1,1 +1,1 @@
-[{"id":1,"genre":"action","author":"sam","source":"youtube","label_id":null,"publish_date":"2023-05-24","publisher":"sam","cover_state":"good"},{"id":12,"genre":"advanture","author":"Sam","source":"Youtube","label_id":null,"publish_date":"2023-05-24","publisher":"sam","cover_state":"good"},{"id":1,"genre":"fiction","author":"sam","source":"youtube","label_id":null,"publish_date":"1999-12-12","publisher":"sams","cover_state":"good"},{"id":12,"genre":"fiction","author":"sam","source":"youtube","label_id":null,"publish_date":"1980-12-24","publisher":"sam","cover_state":"bad"},{"id":45,"genre":"action ","author":"Sam","source":"youtube","label_id":null,"publish_date":"2013-04-13","publisher":"sam","cover_state":"good"},{"id":56,"genre":"fiction ","author":"Sam","source":"Youtube","label_id":4,"publish_date":"1982-02-12","publisher":"Sams","cover_state":"good"},{"id":23,"genre":"","author":"fa","source":"fda","label_id":3,"publish_date":"2009-12-13","publisher":"afdasd","cover_state":"goof"},{"id":0,"genre":"fasdf","author":"fdasf","source":"fdadfsa","label_id":99,"publish_date":"1920-12-12","publisher":"sdfa","cover_state":"bad"}]
+[{"id":1,"genre":null,"author":"sam","source":"youtube","label_id":null,"publish_date":"2023-05-24","publisher":"sam","cover_state":"good"},{"id":12,"genre":null,"author":"Sam","source":"Youtube","label_id":null,"publish_date":"2023-05-24","publisher":"sam","cover_state":"good"},{"id":1,"genre":null,"author":"sam","source":"youtube","label_id":null,"publish_date":"1999-12-12","publisher":"sams","cover_state":"good"},{"id":12,"genre":null,"author":"sam","source":"youtube","label_id":null,"publish_date":"1980-12-24","publisher":"sam","cover_state":"bad"},{"id":45,"genre":null,"author":"Sam","source":"youtube","label_id":null,"publish_date":"2013-04-13","publisher":"sam","cover_state":"good"},{"id":56,"genre":null,"author":"Sam","source":"Youtube","label_id":null,"publish_date":"1982-02-12","publisher":"Sams","cover_state":"good"},{"id":23,"genre":null,"author":"fa","source":"fda","label_id":null,"publish_date":"2009-12-13","publisher":"afdasd","cover_state":"goof"},{"id":0,"genre":null,"author":"fdasf","source":"fdadfsa","label_id":null,"publish_date":"1920-12-12","publisher":"sdfa","cover_state":"bad"},{"id":67,"genre":null,"author":"kimani","source":"sss","label_id":12,"publish_date":"2022-05-08","publisher":"lincon","cover_state":"good"},{"id":323,"genre":null,"author":"efefef","source":"ffef","label_id":4,"publish_date":"2022-04-05","publisher":"ddaad","cover_state":"good"}]
