+++ item.rb
@@ -1,13 +1,14 @@
 class Item
   attr_accessor :id, :genre, :author, :source, :label, :publish_date, :archived
 
-  def initialize(id = Random.rand(1..1000))
+
+  def initialize(id=Random.rand(1..1000), genre='nil', author='', source='', label=1, publish_date ='200-2-2')
     @id = id
-    @genre = nil
-    @author = nil
-    @source = nil
-    @label = nil
-    @publish_date = nil
+    # @genre = nil
+    @author = author
+    @source = source
+    @label = label
+    @publish_date = publish_date
     @archived = false
   end
