+++ labels.json
@@ -1,1 +1,1 @@
-[{"id":99,"name":"Test Label","items":[]},{"id":99,"name":"Test Label","items":[0]},{"id":3,"name":"Hollywood","items":[]},{"id":3,"name":"Nollywood","items":[23]},{"id":4,"name":"Marvel","items":[]},{"id":12,"name":"label","items":[]},{"id":1,"name":"dasd","items":[]}]
+[{"id":99,"name":"Test Label","items":[]},{"id":99,"name":"Test Label","items":[]},{"id":3,"name":"Hollywood","items":[]},{"id":3,"name":"Nollywood","items":[]},{"id":4,"name":"Marvel","items":[323]},{"id":12,"name":"label","items":[]},{"id":1,"name":"dasd","items":[]}]
