--- main1.rb
@@ -1,8 +1,0 @@
-require_relative 'app'
-
-class Main
-  def initialize
-    app = App.new
-    app.main_menu
-  end
-end
